refactor(bench): extract prompt module from bench_serving
Move prompt resolution — truncate_preview, synthetic_prompt_tokens,
synthetic_random_prompt (+ token-id bounds), PromptSpec, resolve_prompt_input,
and the SYNTHETIC_PATTERN label — into prompt.rs. Items main.rs still uses
are pub(crate); trimmed the now-unused rand::RngExt import from main.rs.

Pure code motion, no behavior change. Verified: default build clean (only
the pre-existing ParallelConfig kimi-gated unused-import warning remains).
