+++ openinfer-server/src/bin/bench_serving/main.rs
@@ -29,7 +29,6 @@ use openinfer_core::engine::{EngineLoadOptions, EpBackend};
 #[cfg(feature = "kimi-k2")]
 use openinfer_core::parallel::ParallelConfig;
 use openinfer_vllm_support::load_tokenizer as load_vllm_tokenizer;
-use rand::RngExt;
 use rand::SeedableRng;
 use rand::rngs::StdRng;
 use tokio::sync::mpsc;
@@ -37,10 +36,12 @@ use vllm_text::tokenizer::DynTokenizer;
 
 mod cli;
 mod metrics;
+mod prompt;
 mod render;
 mod report;
 use cli::*;
 use metrics::*;
+use prompt::*;
 use render::*;
 use report::*;
 
@@ -64,118 +65,6 @@ const REGRESSION_TTFT_PCT: f64 = 3.0;
 
 const DEFAULT_REQUEST_PROMPT: &str = "Tell me a story";
 const DEFAULT_CURVE_PROMPT_LEN: usize = 512;
-const SYNTHETIC_PATTERN: &str = "token_id = 100 + (idx % 1000)";
-
-fn truncate_preview(text: &str, limit: usize) -> String {
-    let one_line = text.replace('\n', "\\n");
-    if one_line.chars().count() <= limit {
-        return one_line;
-    }
-    let mut truncated = String::new();
-    for ch in one_line.chars().take(limit) {
-        truncated.push(ch);
-    }
-    truncated.push_str("...");
-    truncated
-}
-
-fn synthetic_prompt_tokens(len: usize) -> Vec<u32> {
-    (0..len).map(|i| ((i % 1000) + 100) as u32).collect()
-}
-
-/// Token-id bounds for synthetic concurrent prompts: above the low special
-/// tokens and well under the smallest supported vocab (DeepSeek-V2-Lite ≈
-/// 102 400), so every drawn id is an ordinary token on any model line.
-const SYNTHETIC_TOKEN_LO: u32 = 100;
-const SYNTHETIC_TOKEN_HI: u32 = 100_000;
-
-/// One synthetic prompt of `len` random tokens, seeded per request so the
-/// concurrent decode streams diverge. Identical concurrent prompts route a MoE
-/// batch onto a narrow expert set, packing the Marlin expert GEMM into fat
-/// tiles and under-measuring decode TPOT by ~7–15% (measured on Kimi-K2 via a
-/// `--distinct-prompts` sweep; the bench trap behind the misread #225 "+51%
-/// HTTP" gap). Distinct prompts exercise realistic broad expert routing. See
-/// docs/lessons/moe-bench-prompt-diversity.md.
-fn synthetic_random_prompt(len: usize, seed: u64, request_idx: usize) -> Vec<u32> {
-    let mut rng =
-        StdRng::seed_from_u64(seed ^ (request_idx as u64).wrapping_mul(0x9E37_79B9_7F4A_7C15));
-    (0..len)
-        .map(|_| rng.random_range(SYNTHETIC_TOKEN_LO..SYNTHETIC_TOKEN_HI))
-        .collect()
-}
-
-#[derive(Debug, Clone)]
-struct PromptSpec {
-    descriptor: PromptDescriptor,
-    tokens: Vec<u32>,
-}
-
-fn resolve_prompt_input(
-    args: &PromptInputArgs,
-    tokenizer: &DynTokenizer,
-    default_text: Option<&str>,
-    default_prompt_len: Option<usize>,
-) -> Result<PromptSpec> {
-    match (&args.prompt, &args.prompt_file, args.prompt_len) {
-        (Some(prompt), None, None) => Ok(PromptSpec {
-            descriptor: PromptDescriptor {
-                source: "text".to_string(),
-                prompt_tokens: tokenizer.encode(prompt, false)?.len(),
-                prompt_preview: Some(truncate_preview(prompt, 96)),
-            },
-            tokens: tokenizer.encode(prompt, false)?,
-        }),
-        (None, Some(path), None) => {
-            let prompt = fs::read_to_string(path)
-                .with_context(|| format!("failed to read prompt file: {path}"))?;
-            let tokens = tokenizer.encode(&prompt, false)?;
-            Ok(PromptSpec {
-                descriptor: PromptDescriptor {
-                    source: format!("file:{path}"),
-                    prompt_tokens: tokens.len(),
-                    prompt_preview: Some(truncate_preview(&prompt, 96)),
-                },
-                tokens,
-            })
-        }
-        (None, None, Some(prompt_len)) => {
-            ensure!(prompt_len > 0, "--prompt-len must be > 0");
-            Ok(PromptSpec {
-                descriptor: PromptDescriptor {
-                    source: format!("synthetic:{SYNTHETIC_PATTERN}"),
-                    prompt_tokens: prompt_len,
-                    prompt_preview: None,
-                },
-                tokens: synthetic_prompt_tokens(prompt_len),
-            })
-        }
-        (None, None, None) => {
-            if let Some(prompt) = default_text {
-                let tokens = tokenizer.encode(prompt, false)?;
-                Ok(PromptSpec {
-                    descriptor: PromptDescriptor {
-                        source: "text".to_string(),
-                        prompt_tokens: tokens.len(),
-                        prompt_preview: Some(truncate_preview(prompt, 96)),
-                    },
-                    tokens,
-                })
-            } else if let Some(prompt_len) = default_prompt_len {
-                Ok(PromptSpec {
-                    descriptor: PromptDescriptor {
-                        source: format!("synthetic:{SYNTHETIC_PATTERN}"),
-                        prompt_tokens: prompt_len,
-                        prompt_preview: None,
-                    },
-                    tokens: synthetic_prompt_tokens(prompt_len),
-                })
-            } else {
-                unreachable!("default prompt source must be provided");
-            }
-        }
-        _ => unreachable!("clap enforces prompt input conflicts"),
-    }
-}
 
 struct GenTimings {
     ttft: Duration,
